+++ common/src/main/java/net/spell_engine/config/ServerConfig.java
@@ -22,7 +22,7 @@ public class ServerConfig implements ConfigData { public ServerConfig() {}
     @Comment("Spells should damage items on use. Set `false` to disable.")
     public boolean spell_cost_durability_allowed = true;
     @Comment("The time in ticks of global cooldown to apply to all instant cast spells when casted.")
-    public int spell_instant_cast_gcd = 0;
+    public int spell_instant_cast_global_cooldown = 10;
     @Comment("Players cannot unequip a spell book, if one of the spells in it is on cooldown.")
     public boolean spell_item_cooldown_lock = true;
     @Comment("Players can use the Spell Binding Table to create spell books.")